feat: adjustBalancesFacet
