+++ .changeset/maf-split-adjust-balances.md
@@ -1,0 +1,5 @@
+---
+"@hashgraph/asset-tokenization-contracts": minor
+---
+
+Consolidate 8 balance-adjustment functions from ScheduledBalanceAdjustmentsFacet, EquityUSAFacet, and ERC1410TokenHolderFacet into a single AdjustBalancesFacet at the flat contracts/facets/adjustBalances/ path.
