+++ packages/ats/contracts/contracts/domain/asset/EquityStorageWrapper.sol
@@ -8,6 +8,7 @@ import {
     KPI_EQUITY_BALANCE_ADJ
 } from "../../constants/values.sol";
 import { IEquity } from "../../facets/layer_2/equity/IEquity.sol";
+import { IAdjustBalances } from "../../facets/adjustBalances/IAdjustBalances.sol";
 import { CorporateActionsStorageWrapper } from "../core/CorporateActionsStorageWrapper.sol";
 import { NominalValueStorageWrapper } from "./nominalValue/NominalValueStorageWrapper.sol";
 import { ScheduledTasksStorageWrapper } from "./ScheduledTasksStorageWrapper.sol";
@@ -56,7 +57,7 @@ library EquityStorageWrapper {
     }
 
     function setScheduledBalanceAdjustment(
-        IEquity.ScheduledBalanceAdjustment calldata newBalanceAdjustment
+        IAdjustBalances.ScheduledBalanceAdjustment calldata newBalanceAdjustment
     ) internal returns (bytes32 corporateActionId_, uint256 balanceAdjustmentID_) {
         bytes memory data = abi.encode(newBalanceAdjustment);
 
@@ -69,11 +70,15 @@ library EquityStorageWrapper {
     }
 
     function cancelScheduledBalanceAdjustment(uint256 balanceAdjustmentId) internal returns (bool success_) {
-        IEquity.ScheduledBalanceAdjustment memory balanceAdjustment;
+        CorporateActionsStorageWrapper.requireMatchingActionType(
+            BALANCE_ADJUSTMENT_CORPORATE_ACTION_TYPE,
+            balanceAdjustmentId - 1
+        );
+        IAdjustBalances.ScheduledBalanceAdjustment memory balanceAdjustment;
         bytes32 corporateActionId;
         (balanceAdjustment, corporateActionId, ) = getScheduledBalanceAdjustment(balanceAdjustmentId);
         if (balanceAdjustment.executionDate <= TimeTravelStorageWrapper.getBlockTimestamp()) {
-            revert IEquity.BalanceAdjustmentAlreadyExecuted(corporateActionId, balanceAdjustmentId);
+            revert IAdjustBalances.BalanceAdjustmentAlreadyExecuted(corporateActionId, balanceAdjustmentId);
         }
         CorporateActionsStorageWrapper.cancelCorporateAction(corporateActionId);
         success_ = true;
@@ -81,12 +86,12 @@ library EquityStorageWrapper {
 
     function initBalanceAdjustment(bytes32 actionId, bytes memory data) internal {
         if (actionId == bytes32(0)) {
-            revert IEquity.BalanceAdjustmentCreationFailed();
+            revert IAdjustBalances.BalanceAdjustmentCreationFailed();
         }
 
-        IEquity.ScheduledBalanceAdjustment memory newBalanceAdjustment = abi.decode(
+        IAdjustBalances.ScheduledBalanceAdjustment memory newBalanceAdjustment = abi.decode(
             data,
-            (IEquity.ScheduledBalanceAdjustment)
+            (IAdjustBalances.ScheduledBalanceAdjustment)
         );
 
         ScheduledTasksStorageWrapper.addScheduledCrossOrderedTask(
@@ -141,7 +146,7 @@ library EquityStorageWrapper {
         internal
         view
         returns (
-            IEquity.ScheduledBalanceAdjustment memory balanceAdjustment_,
+            IAdjustBalances.ScheduledBalanceAdjustment memory balanceAdjustment_,
             bytes32 corporateActionId_,
             bool isDisabled_
         )
@@ -155,7 +160,7 @@ library EquityStorageWrapper {
         (, , data, isDisabled_) = CorporateActionsStorageWrapper.getCorporateAction(corporateActionId_);
 
         _checkUnexpectedError(data.length == 0, KPI_EQUITY_BALANCE_ADJ);
-        (balanceAdjustment_) = abi.decode(data, (IEquity.ScheduledBalanceAdjustment));
+        (balanceAdjustment_) = abi.decode(data, (IAdjustBalances.ScheduledBalanceAdjustment));
     }
 
     function getScheduledBalanceAdjustmentsCount() internal view returns (uint256 balanceAdjustmentCount_) {
