--- packages/ats/contracts/contracts/facets/layer_2/adjustBalance/AdjustBalancesFacet.sol
@@ -1,25 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity >=0.8.0 <0.9.0;
-
-import { IAdjustBalances } from "./IAdjustBalances.sol";
-import { AdjustBalances } from "./AdjustBalances.sol";
-import { IStaticFunctionSelectors } from "../../../infrastructure/proxy/IStaticFunctionSelectors.sol";
-import { _BALANCE_ADJUSTMENTS_RESOLVER_KEY } from "../../../constants/resolverKeys.sol";
-
-contract AdjustBalancesFacet is AdjustBalances, IStaticFunctionSelectors {
-    function getStaticResolverKey() external pure override returns (bytes32 staticResolverKey_) {
-        staticResolverKey_ = _BALANCE_ADJUSTMENTS_RESOLVER_KEY;
-    }
-
-    function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
-        uint256 selectorIndex;
-        staticFunctionSelectors_ = new bytes4[](1);
-        staticFunctionSelectors_[selectorIndex++] = this.adjustBalances.selector;
-    }
-
-    function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
-        staticInterfaceIds_ = new bytes4[](1);
-        uint256 selectorsIndex;
-        staticInterfaceIds_[selectorsIndex++] = type(IAdjustBalances).interfaceId;
-    }
-}
