--- packages/ats/contracts/contracts/facets/layer_2/adjustBalance/IAdjustBalances.sol
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity >=0.8.0 <0.9.0;
-
-interface IAdjustBalances {
-    event AdjustmentBalanceSet(address indexed operator, uint256 factor, uint8 decimals);
-
-    error FactorIsZero();
-
-    /**
-     * @notice Adjusts the balances of all users by a given factor and decimals
-     * @dev This action is triggered inmediately, contrary to the secheduled methods that add tasks to the queue
-     */
-    function adjustBalances(uint256 factor, uint8 decimals) external returns (bool success_);
-}
