+++ packages/ats/contracts/contracts/infrastructure/utils/DefaultValueValidation.sol
@@ -4,10 +4,10 @@ pragma solidity >=0.8.0 <0.9.0;
 import { ICommonErrors } from "../errors/ICommonErrors.sol";
 
 /**
- * @title Address Validation
- * @notice Utility library for address validation
+ * @title Default Value Validation
+ * @notice Utility library for validating that values are not the Solidity default (zero)
  */
-library AddressValidation {
+library DefaultValueValidation {
     /**
      * @notice Validates that the provided address is not the zero address
      * @param _address The address to validate
@@ -15,6 +15,16 @@ library AddressValidation {
     function checkZeroAddress(address _address) internal pure {
         if (_address == address(0)) {
             revert ICommonErrors.ZeroAddressNotAllowed();
+        }
+    }
+
+    /**
+     * @notice Validates that the provided uint256 is not zero
+     * @param _value The value to validate
+     */
+    function checkZeroValue(uint256 _value) internal pure {
+        if (_value == 0) {
+            revert ICommonErrors.ZeroValueNotAllowed();
         }
     }
 }
