+++ packages/ats/contracts/contracts/services/core/CoreModifiers.sol
@@ -21,7 +21,7 @@ pragma solidity >=0.8.0 <0.9.0;
  */
 
 import { AccessControlModifiers } from "./AccessControlModifiers.sol";
-import { AddressModifiers } from "./AddressModifiers.sol";
+import { DefaultValuesModifiers } from "./DefaultValuesModifiers.sol";
 import { CapModifiers } from "./CapModifiers.sol";
 import { ControlListModifiers } from "./ControlListModifiers.sol";
 import { DateValidationModifiers } from "./DateValidationModifiers.sol";
@@ -36,7 +36,7 @@ import { DeactivateModifiers } from "./DeactivateModifiers.sol";
 
 abstract contract CoreModifiers is
     AccessControlModifiers,
-    AddressModifiers,
+    DefaultValuesModifiers,
     CapModifiers,
     ControlListModifiers,
     DateValidationModifiers,
