+++ packages/ats/contracts/scripts/domain/bond/createConfiguration.ts
@@ -124,7 +124,6 @@ const BOND_FACETS = [
   "NominalValueFacet",
   "ProceedRecipientsFacet",
   "ProtectedPartitionsFacet",
-  "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
   "CouponListingFacet",
   "SsiManagementFacet",
