+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -117,7 +117,6 @@ const BOND_FIXED_RATE_FACETS = [
   "NominalValueFacet",
   "ProceedRecipientsFacet",
   "ProtectedPartitionsFacet",
-  "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
   "CouponListingFacet",
   "SsiManagementFacet",
