+++ packages/ats/contracts/scripts/domain/equity/createConfiguration.ts
@@ -114,7 +114,6 @@ const EQUITY_FACETS = [
   "LockFacet",
   "NominalValueFacet",
   "ProtectedPartitionsFacet",
-  "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
   "SsiManagementFacet",
   "TransferAndLockFacet",
