+++ packages/ats/contracts/scripts/domain/loan/createConfiguration.ts
@@ -111,7 +111,6 @@ const LOAN_FACETS = [
   "ClearingByPartitionFacet",
 
   // Scheduled Tasks
-  "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
 
   // External Management
