+++ packages/ats/contracts/scripts/domain/loanPortfolio/createConfiguration.ts
@@ -114,7 +114,6 @@ const LOANS_PORTFOLIO_FACETS = [
   "ClearingByPartitionFacet",
 
   // Scheduled Tasks
-  "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
   "CouponListingFacet",
