+++ packages/ats/contracts/test/contracts/integration/layer_1/equity/equity.test.ts
@@ -1010,208 +1010,4 @@ describe("Equity Tests", () => {
       });
     });
   });
-
-  describe("Balance adjustments", () => {
-    it("GIVEN an account without corporateActions role WHEN setScheduledBalanceAdjustment THEN transaction fails with AccountHasNoRole", async () => {
-      // set dividend fails
-      await expect(
-        asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData),
-      ).to.be.revertedWithCustomError(asset, "AccountHasNoRole");
-    });
-
-    it("GIVEN a paused Token WHEN setScheduledBalanceAdjustment THEN transaction fails with TokenIsPaused", async () => {
-      // Granting Role to account C and Pause
-      await grantRoleAndPauseToken(asset, ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A, signer_B, signer_C.address);
-
-      // set dividend fails
-      await expect(
-        asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData),
-      ).to.be.revertedWithCustomError(asset, "TokenIsPaused");
-    });
-
-    it("GIVEN an account with corporateActions role WHEN setScheduledBalanceAdjustment with invalid timestamp THEN transaction fails with WrongTimestamp", async () => {
-      const currentTimestamp = await asset.blockTimestamp();
-      await asset.changeSystemTimestamp(currentTimestamp + 100n);
-      await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-      const invalidBalanceAdjustmentData = {
-        executionDate: (currentTimestamp - 100n).toString(), // Past timestamp
-        factor: balanceAdjustmentFactor,
-        decimals: balanceAdjustmentDecimals,
-      };
-
-      await expect(
-        asset.connect(signer_C).setScheduledBalanceAdjustment(invalidBalanceAdjustmentData),
-      ).to.be.revertedWithCustomError(asset, "WrongTimestamp");
-    });
-
-    it("GIVEN an account with corporateActions role WHEN setScheduledBalanceAdjustment with invalid factor THEN transaction fails with FactorIsZero", async () => {
-      await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-      const invalidBalanceAdjustmentData = {
-        executionDate: balanceAdjustmentExecutionDateInSeconds.toString(),
-        factor: 0, // Invalid factor: 0
-        decimals: balanceAdjustmentDecimals,
-      };
-
-      await expect(
-        asset.connect(signer_C).setScheduledBalanceAdjustment(invalidBalanceAdjustmentData),
-      ).to.be.revertedWithCustomError(asset, "FactorIsZero");
-    });
-
-    it("GIVEN balance adjustment created WHEN trying to get balance adjustment with wrong ID type THEN transaction fails with WrongIndexForAction", async () => {
-      await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-      // Create a balance adjustment
-      await asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData);
-
-      // Create a dividend to have different action types
-      await asset.connect(signer_C).setDividend(dividendData);
-
-      // Try to access balance adjustment with dividend ID (should fail)
-      await expect(asset.getScheduledBalanceAdjustment(2)).to.be.revertedWithCustomError(asset, "WrongIndexForAction");
-    });
-
-    it("GIVEN an account with corporateActions role WHEN setScheduledBalanceAdjustment THEN transaction succeeds", async () => {
-      // Granting Role to account C
-      await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-      // set dividend
-      await expect(asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData))
-        .to.emit(asset, "ScheduledBalanceAdjustmentSet")
-        .withArgs(
-          "0x0000000000000000000000000000000000000000000000000000000000000001",
-          1,
-          signer_C.address,
-          balanceAdjustmentExecutionDateInSeconds,
-          balanceAdjustmentFactor,
-          balanceAdjustmentDecimals,
-        );
-
-      const listCount = await asset.getScheduledBalanceAdjustmentCount();
-      const [balanceAdjustment, isDisabled] = await asset.getScheduledBalanceAdjustment(1);
-
-      expect(listCount).to.equal(1);
-      expect(isDisabled).to.equal(false);
-      expect(balanceAdjustment.executionDate).to.equal(balanceAdjustmentExecutionDateInSeconds);
-      expect(balanceAdjustment.factor).to.equal(balanceAdjustmentFactor);
-      expect(balanceAdjustment.decimals).to.equal(balanceAdjustmentDecimals);
-    });
-
-    describe("Cancel Scheduled Balance Adjustment", () => {
-      it("GIVEN an account without corporateActions role WHEN cancelScheduledBalanceAdjustment THEN transaction fails with AccountHasNoRole", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_B.address);
-        await asset.connect(signer_B).setScheduledBalanceAdjustment(balanceAdjustmentData);
-        await expect(asset.connect(signer_C).cancelScheduledBalanceAdjustment(1)).to.be.revertedWithCustomError(
-          asset,
-          "AccountHasNoRole",
-        );
-      });
-
-      it("GIVEN a paused Token WHEN cancelScheduledBalanceAdjustment THEN transaction fails with TokenIsPaused", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_B.address);
-        await asset.connect(signer_B).setScheduledBalanceAdjustment(balanceAdjustmentData);
-        await asset.connect(signer_B).pause();
-
-        await expect(asset.connect(signer_B).cancelScheduledBalanceAdjustment(1)).to.be.revertedWithCustomError(
-          asset,
-          "TokenIsPaused",
-        );
-      });
-
-      it("GIVEN a balance adjustment already executed WHEN cancelScheduledBalanceAdjustment THEN transaction fails with BalanceAdjustmentAlreadyExecuted", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-        await asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData);
-
-        await asset.changeSystemTimestamp(balanceAdjustmentExecutionDateInSeconds + 1000);
-
-        await expect(asset.connect(signer_C).cancelScheduledBalanceAdjustment(1)).to.be.revertedWithCustomError(
-          asset,
-          "BalanceAdjustmentAlreadyExecuted",
-        );
-      });
-
-      it("GIVEN a balance adjustment not yet executed WHEN cancelScheduledBalanceAdjustment THEN transaction succeeds", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-        await asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData);
-
-        await expect(asset.connect(signer_C).cancelScheduledBalanceAdjustment(1))
-          .to.emit(asset, "ScheduledBalanceAdjustmentCancelled")
-          .withArgs(1, signer_C.address);
-        const [balanceAdjustment, isDisabled] = await asset.getScheduledBalanceAdjustment(1);
-        expect(isDisabled).to.equal(true);
-        expect(balanceAdjustment.factor).to.equal(balanceAdjustmentFactor);
-        expect(balanceAdjustment.decimals).to.equal(balanceAdjustmentDecimals);
-      });
-
-      it("GIVEN a non-existent balance adjustment WHEN cancelScheduledBalanceAdjustment THEN transaction fails", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-        await expect(asset.connect(signer_C).cancelScheduledBalanceAdjustment(999)).to.be.rejected;
-      });
-
-      it("GIVEN multiple balance adjustments WHEN cancelScheduledBalanceAdjustment on one THEN only that adjustment is cancelled", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-
-        // Create first balance adjustment
-        await asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData);
-
-        // Create second balance adjustment with different execution date
-        const secondBalanceAdjustmentData = {
-          executionDate: (balanceAdjustmentExecutionDateInSeconds + 10000).toString(),
-          factor: 500,
-          decimals: 3,
-        };
-        await asset.connect(signer_C).setScheduledBalanceAdjustment(secondBalanceAdjustmentData);
-
-        // Cancel only first balance adjustment
-        await expect(asset.connect(signer_C).cancelScheduledBalanceAdjustment(1))
-          .to.emit(asset, "ScheduledBalanceAdjustmentCancelled")
-          .withArgs(1, signer_C.address);
-
-        // Verify first adjustment is cancelled by checking it still exists but with correct data
-        const [balanceAdjustment1, isDisabled1] = await asset.getScheduledBalanceAdjustment(1);
-        expect(isDisabled1).to.equal(true);
-        expect(balanceAdjustment1.factor).to.equal(balanceAdjustmentFactor);
-
-        // Verify second adjustment is still active
-        const [balanceAdjustment2, isDisabled2] = await asset.getScheduledBalanceAdjustment(2);
-        expect(isDisabled2).to.equal(false);
-        expect(balanceAdjustment2.factor).to.equal(500);
-        expect(balanceAdjustment2.decimals).to.equal(3);
-      });
-
-      it("GIVEN a cancelled balance adjustment WHEN triggerScheduledCrossOrderedTasks is called THEN scheduled task executes but token balance remains unchanged", async () => {
-        await asset.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_C.address);
-        await asset.connect(signer_A).grantRole(ATS_ROLES.ISSUER_ROLE, signer_C.address);
-
-        const tokenAmount = 1000n;
-        await asset.connect(signer_C).issueByPartition({
-          partition: DEFAULT_PARTITION,
-          tokenHolder: signer_A.address,
-          value: tokenAmount,
-          data: "0x",
-        });
-
-        const balanceBeforeAdjustment = await asset.balanceOfByPartition(DEFAULT_PARTITION, signer_A.address);
-        expect(balanceBeforeAdjustment).to.equal(tokenAmount);
-
-        await asset.connect(signer_C).setScheduledBalanceAdjustment(balanceAdjustmentData);
-
-        await asset.connect(signer_C).cancelScheduledBalanceAdjustment(1);
-
-        const balanceAfterCancel = await asset.balanceOfByPartition(DEFAULT_PARTITION, signer_A.address);
-        expect(balanceAfterCancel).to.equal(balanceBeforeAdjustment);
-
-        await asset.changeSystemTimestamp(balanceAdjustmentExecutionDateInSeconds + 1);
-
-        await asset.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
-
-        const balanceAfterTrigger = await asset.balanceOfByPartition(DEFAULT_PARTITION, signer_A.address);
-        expect(balanceAfterTrigger).to.equal(balanceBeforeAdjustment);
-      });
-    });
-  });
 });
